refactor (WinningLotto) : 로또 당첨 여부 확인의 책임을 WinningLotto로 리팩토링 한다
