+++ src/test/java/lotto/domain/LottoTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayName("Lotto 클래스")
@@ -31,15 +30,5 @@ class LottoTest {
     void createLottoByNull() {
         assertThatThrownBy(() -> new Lotto(null))
             .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("로또 번호와 당첨 번호를 비교하여 당첨 여부를 판단한다.")
-    @Test
-    void compareLottoNumbersToWinningNumbers() {
-        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        WinningLotto winningLotto = new WinningLotto(
-            List.of(1, 2, 3, 4, 5, 6), new BonusNumber(7));
-
-        assertThat(lotto.matches(winningLotto)).isEqualTo(LottoPrize.FIRST);
     }
 }
